move checker package out of internal and expose it
