+++ checker/checker.go
